// libs/sq-server-commons/src/types/browser.ts
/*
 * SonarQube
 * Copyright (C) 2009-2025 SonarSource Sàrl
 * mailto:info AT sonarsource DOT com
 *
 * This program is free software; you can redistribute it and/or
 * modify it under the terms of the GNU Lesser General Public
 * License as published by the Free Software Foundation; either
 * version 3 of the License, or (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
 * Lesser General Public License for more details.
 *
 * You should have received a copy of the GNU Lesser General Public License
 * along with this program; if not, write to the Free Software Foundation,
 * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 */

import { ExtensionStartMethod } from './extension';
import { InstanceType } from './system';
import { SysStatus } from './types';

export interface EnhancedWindow extends Window {
  baseUrl: string;
  instance: InstanceType;
  official: boolean;
  registerExtension: (key: string, start: ExtensionStartMethod, providesCSSFile?: boolean) => void;

  serverStatus: SysStatus;
  setWebAnalyticsPageChangeHandler: (pageHandler: (pathname: string) => void) => void;
  t: (...keys: string[]) => string;
  tp: (messageKey: string, ...parameters: Array<string | number>) => string;
}

export interface AppVariablesElement extends HTMLElement {
  dataset: {
    baseUrl: string;
    instance: InstanceType;
    official: string;
    serverStatus: SysStatus;
  };
}